Update = Delete and Edit

+++ src/main/java/com/bootcamp/techno/model/SkripsiModel.java
@@ -1,5 +1,7 @@
 package com.bootcamp.techno.model;
 
+import java.sql.Date;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -11,14 +13,16 @@ import lombok.Data;
 @Data
 @Entity
 @Table(name = "T_SKRIPSI")
+
 public class SkripsiModel {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int id;
 	private String judul;
-	private int nilai;
+	private double nilai;
 	private int tahun;
+	
 	public int getId() {
 		return id;
 	}
@@ -31,10 +35,10 @@ public class SkripsiModel {
 	public void setJudul(String judul) {
 		this.judul = judul;
 	}
-	public int getNilai() {
+	public double getNilai() {
 		return nilai;
 	}
-	public void setNilai(int nilai) {
+	public void setNilai(double nilai) {
 		this.nilai = nilai;
 	}
 	public int getTahun() {
